% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management_defaults.R
\name{add_management}
\alias{add_management}
\title{Estimate default values for management}
\usage{
add_management(
  ID,
  B_LU_BRP,
  B_SOILTYPE_AGR,
  M_GREEN = NA,
  M_NONBARE = NA,
  M_EARLYCROP = NA,
  M_COMPOST = NA,
  M_SLEEPHOSE = NA,
  M_DRAIN = NA,
  M_DITCH = NA,
  M_UNDERSEED = NA
)
}
\arguments{
\item{ID}{(character) A field id}

\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop (optional, option: yes or no)}

\item{M_NONBARE}{(boolean) A soil measure. Is parcel for 80 percent of the year cultivated and 'green' (optional, option: yes or no)}

\item{M_EARLYCROP}{(boolean) A soil measure. Use of early crop varieties to avoid late harvesting (optional, option: yes or no)}

\item{M_COMPOST}{(numeric) The frequency that compost is applied (optional, every x years)}

\item{M_SLEEPHOSE}{(boolean) A soil measure. Is sleephose used for slurry application (optional, option: yes or no)}

\item{M_DRAIN}{(boolean) A soil measure. Are under water drains installed in peaty soils (optional, option: yes or no)}

\item{M_DITCH}{(boolean) A soil measure. Are ditched maintained carefully and slib applied on the land (optional, option: yes or no)}

\item{M_UNDERSEED}{(boolean) A soil measure. Is grass used as second crop in between maize rows (optional, option: yes or no)}
}
\description{
This function adds default management input variables given soil type and land use
}
