% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterretention.R
\docType{data}
\name{bouwsteen_tb}
\alias{bouwsteen_tb}
\title{Table with water retention properties of 'bouwstenen'}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 36 rows and 14 columns.}
\usage{
bouwsteen_tb
}
\description{
This table contains water retention curve parameters and typical mineral composition of 18 bouwstenen
}
\keyword{datasets}
