% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{calc_aggregatestability}
\alias{calc_aggregatestability}
\title{Calculate aggregate stability index based on occupation CEC}
\usage{
calc_aggregatestability(B_BT_AK, A_OS_GV, A_K_CEC, A_CA_CEC, A_MG_CEC)
}
\arguments{
\item{B_BT_AK}{(character) The type of soil}

\item{A_OS_GV}{(numeric) The organic matter content of soil in percentage}

\item{A_K_CEC}{(numeric) The occupation of the CEC with K (\%)}

\item{A_CA_CEC}{(numeric) The occupation of the CEC with Mg (\%)}

\item{A_MG_CEC}{(numeric) The occupation of the CEC with Ca (\%)}
}
\description{
This function calculates an aggregate stability index given the CEC and its occupation with major cations.
}
