% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_density.R
\name{calc_bulk_density}
\alias{calc_bulk_density}
\title{Calculate the bulk density}
\usage{
calc_bulk_density(A_OS_GV, B_BT_AK, A_CLAY_MI = NULL)
}
\arguments{
\item{A_OS_GV}{(numeric) The percentage organic matter in the soil}

\item{B_BT_AK}{(character) The agricultural type of soil}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}
}
\description{
This function calculates the bulk density of the soil based on texture and organic matter
}
