% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_model.R
\name{calc_c_seq_field}
\alias{calc_c_seq_field}
\title{Evaluate the carbon sequestration rate}
\usage{
calc_c_seq_field(
  B_LU_BRP,
  B_SOILTYPE_AGR,
  A_SOM_LOI,
  A_CLAY_MI,
  A_P_AL,
  A_P_WA,
  A_DEPTH = 0.3,
  A_TEMP_MEAN = NULL,
  A_PREC_MEAN = NULL,
  A_ET_MEAN = NULL,
  M_GREEN = FALSE,
  effectivity = TRUE,
  manure_in = NULL,
  compost_in = 0,
  manure_type = "slurry",
  history = "default",
  renewal = NULL,
  grass_fertilization = 1,
  c_fractions = c(0.0558, 0.015, 0.125, 0.015),
  dec_rates = c(10, 0.3, 0.66, 0.02)
)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_P_AL}{(numeric) The P-AL content of the soil}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water (mg P2O5 / 100 ml soil)}

\item{A_DEPTH}{(numeric) Depth of the soil layer (m)}

\item{A_TEMP_MEAN}{(numeric) Mean monthly temperature (dC), should be a vector of 12 elements, optional}

\item{A_PREC_MEAN}{(numeric) Mean monthly precipitation (mm), should be a vector of 12 elements, optional}

\item{A_ET_MEAN}{(numeric) Mean actual evapo-transpiration (mm), should be a vector of 12 elements, optional}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop, optional}

\item{effectivity}{(boolean) A vector that tells whether the catch crop was effective (i.e. did it grow sufficiently), optional}

\item{manure_in}{(numeric) Annual amount of C applied to the soil via manure (kg C/ha), should be a vector with a value per year, optional}

\item{compost_in}{(numeric) Annual amount of C applied to the soil via compost (kg C/ha), should be a vector with a value per year, optional}

\item{manure_type}{(character) The type of manure applied on the field, options: 'slurry' or 'solid', should be a vector with a value per year, optional}

\item{history}{(character) The manure history of the soil, optional (options: 'default', 'grass_renewal' for grassland renewal, 'manure' for intensive manure application and 'manual')}

\item{renewal}{(numeric) The years in which grassland renewal takes place (vector of years), optional}

\item{grass_fertilization}{(numeric) Timing of fertilization application, optional (options: 1, 2, 3)}

\item{c_fractions}{(numeric) A vector of the fractions of total carbon in the IOM, DPM, RPM and BIO pool (-), the fraction of the HUM pool is derived form these values; if not provided, default values are used, optional}

\item{dec_rates}{(numeric) A vector of the decomposition rate constants for the DPM, RPM, BIO and HUM pool (/year); if not provided, default values are used, optional}
}
\description{
This function is a wrapper function that calculates the evaluation of the C pools over time for the different management types and produces a score for carbon sequestration
}
\details{
The output of this function is a list containing the data.tables with time and OM levels for current, optimal and minimal management, respectively. The last item in the list is the index score for carbon sequestration.

The input parameters M_GREEN, effectivity, manure_in, compost_in and grass_fertilization should contain a vector with a value for each year of the supplied crop rotation in B_LU_BRP. If only one value is provided, this value will be assigned to each year.
}
