% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_model.R
\name{calc_carbon_input}
\alias{calc_carbon_input}
\title{Calculate the carbon inputs on a field}
\usage{
calc_carbon_input(
  B_LU_BRP,
  A_P_AL,
  A_P_WA,
  M_GREEN = FALSE,
  effectivity = TRUE,
  manure_type = "slurry",
  manure_in = NULL,
  compost_in = NULL
)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{A_P_AL}{(numeric) The P-AL content of the soil}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water (mg P2O5 / 100 ml soil)}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop, optional}

\item{effectivity}{(boolean) A vector that tells whether the catch crop was effective (i.e. did it grow sufficiently), optional}

\item{manure_type}{(character) The type of manure applied on the field, options: 'slurry' or 'solid', should be a vector with a value per year}

\item{manure_in}{(numeric) Annual amount of C applied to the soil via manure (kg C/ha), should be a vector with a value per year, optional, if NU}

\item{compost_in}{(numeric) Annual amount of C applied to the soil via compost (kg C/ha), should be a vector with a value per year, optional}
}
\description{
This function calculates the carbon inputs to the field based on manure type, P status of the soil and management practices
}
