% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rothc.R
\name{calc_cor_factors}
\alias{calc_cor_factors}
\title{Calculate the correction factors for the RothC model}
\usage{
calc_cor_factors(
  A_TEMP_MEAN,
  A_PREC_MEAN,
  A_ET_MEAN,
  A_CLAY_MI,
  A_DEPTH = 0.3,
  crop_cover,
  mcf,
  renewal = NULL,
  simyears
)
}
\arguments{
\item{A_TEMP_MEAN}{(numeric) Mean monthly temperature (dC), should be a vector of 12 elements, optional}

\item{A_PREC_MEAN}{(numeric) Mean monthly precipitation (mm), should be a vector of 12 elements, optional}

\item{A_ET_MEAN}{(numeric) Mean actual evapo-transpiration (mm), should be a vector of 12 elements, optional}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_DEPTH}{(numeric) Depth of the soil layer (m)}

\item{crop_cover}{(numeric) A vector with crop cover of the soil per month (options: 1 or 0)}

\item{mcf}{(numeric) A vector with Makkink correction factor for evapo-transpiration per month (between 0 and 2)}

\item{renewal}{(numeric) A vector with the years in which grassland renewal takes place, optional}

\item{simyears}{(numeric) Amount of years for which the simulation should run}
}
\description{
This function calculates the correction factors for the RothC model
}
\references{
Coleman & Jenkinson (1996) RothC - A model for the turnover of carbon in soil
}
