% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rothc.R
\name{calc_cpools}
\alias{calc_cpools}
\title{Initialization of the RothC carbon pools}
\usage{
calc_cpools(
  B_SOILTYPE_AGR,
  A_SOM_LOI,
  A_CLAY_MI,
  A_DEPTH = 0.3,
  history = "default",
  c_fractions
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_DEPTH}{(numeric) Depth of the soil layer (m)}

\item{history}{(character) The manure history of the soil, optional (options: default, grass_rn for grassland renewal, manure for intensive manure application and manual)}

\item{c_fractions}{(numeric) A vector of the fractions of total carbon in the IOM, DPM, RPM and BIO pool (-), the fraction of the HUM pool is derived form these values; if not provided, default values are used, optional}
}
\description{
This function calculates the initial size of the carbon pools of the RothC model
}
\references{
Coleman & Jenkinson (1996) RothC - A model for the turnover of carbon in soil
}
