% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{calc_depth}
\alias{calc_depth}
\title{Determine the depth of the soil}
\usage{
calc_depth(B_LU_BRP)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}
}
\description{
This function determines the depth of the soil
}
