% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundwater_recharge.R
\name{calc_gw_recharge}
\alias{calc_gw_recharge}
\title{Calculate groundwater recharge through precipitation surplus in the Netherlands}
\usage{
calc_gw_recharge(ID, B_LU_BRP, M_GREEN)
}
\arguments{
\item{ID}{(numeric) The ID of the field}

\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop (optional, option: yes or no)}
}
\description{
This function calculates the volume of water (in mm / ha) that infiltrates to the groundwater given the crop rotation plan.
}
