% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potassium.R
\name{calc_k_availability}
\alias{calc_k_availability}
\title{Calculate the K availability}
\usage{
calc_k_availability(A_K_CC, A_K_CEC, A_CEC_CO, A_PH_CC, A_OS_GV, A_CLAY_MI,
  B_LU_BRP, B_BT_AK)
}
\arguments{
\item{A_K_CC}{(numeric) The K-CaCl2 content of the soil in mg K / kg}

\item{A_CEC_CO}{(numeric) The CEC content of the soil (mmol+ per kg)}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}
}
\description{
This function calculates the K availability of a soil.
}
