% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{calc_management}
\alias{calc_management}
\title{Calculate the 'performance' of sustainable soil management}
\usage{
calc_management(
  A_SOM_LOI,
  B_LU_BRP,
  B_SOILTYPE_AGR,
  B_GWL_CLASS,
  D_SOM_BAL,
  D_CP_GRASS,
  D_CP_POTATO,
  D_CP_RUST,
  D_CP_RUSTDEEP,
  D_GA,
  M_GREEN,
  M_NONBARE,
  M_EARLYCROP,
  M_SLEEPHOSE,
  M_DRAIN,
  M_DITCH,
  M_UNDERSEED
)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{D_SOM_BAL}{(numeric) The organic matter balance of the soil (in kg EOS per ha)}

\item{D_CP_GRASS}{(numeric) The fraction grassland in crop rotation}

\item{D_CP_POTATO}{(numeric) The fraction potato crops in crop rotation}

\item{D_CP_RUST}{(numeric) The fraction rustgewassen in crop rotation}

\item{D_CP_RUSTDEEP}{(numeric) The fraction diepe rustgewassen in crop rotation (-)}

\item{D_GA}{(numeric) The age of the grassland (years)}

\item{M_GREEN}{(boolean) measure. are catchcrops sown after main crop (option: yes or no)}

\item{M_NONBARE}{(boolean) measure. is parcel for 80 percent of the year cultivated and 'green' (option: yes or no)}

\item{M_EARLYCROP}{(boolean) measure. use of early crop varieties to avoid late harvesting (option: yes or no)}

\item{M_SLEEPHOSE}{(boolean) measure. is sleepslangbemester used for slurry application (option: yes or no)}

\item{M_DRAIN}{(boolean) measure. are under water drains installed in peaty soils (option: yes or no)}

\item{M_DITCH}{(boolean) measure. are ditched maintained carefully and slib applied on the land (option: yes or no)}

\item{M_UNDERSEED}{(boolean) measure. is grass used as second crop in between maize rows (option: yes or no)}
}
\description{
This function evaluates the contribution of sustainable soil management following the Label Sustainable Soil Management.
}
