% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nretention.r
\name{calc_nleach}
\alias{calc_nleach}
\title{Calculate the N leaching}
\usage{
calc_nleach(B_BT_AK, B_LU_BRP, B_GT, D_NLV, wtype)
}
\arguments{
\item{B_BT_AK}{(character) The type of soil}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}

\item{B_GT}{(character) The groundwater table class}

\item{D_NLV}{(numeric) The N-leverend vermogen (kg N ha-1 jr-1) calculated by \code{\link{calc_nlv}}}

\item{wtype}{(character) whether it computes N leaching to groundwater ("gw") or to surface water ("ow")}
}
\description{
This function calculates the potential N leaching of a soil.
}
