% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organic_carbon.R
\name{calc_organic_carbon}
\alias{calc_organic_carbon}
\title{Calculate amount of organic carbon}
\usage{
calc_organic_carbon(A_OS_GV, D_BDS, D_RD)
}
\arguments{
\item{A_OS_GV}{(numeric) The percentage organic matter in the soil}

\item{D_BDS}{(numeric) The bulk density of the soil}

\item{D_RD}{(numeric) The root depth of the crop}
}
\description{
This function calculates the amount of organic carbon in the soil
}
