% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundwater_recharge.R
\name{calc_permeability}
\alias{calc_permeability}
\title{Calculate the permeability of the top soil}
\usage{
calc_permeability(A_CLAY_MI, A_SAND_MI, A_SILT_MI, A_SOM_LOI)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}
}
\description{
This function calculates the permeability of the top soil
}
