% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{calc_ph_delta}
\alias{calc_ph_delta}
\title{Calculate the difference between pH and optimum}
\usage{
calc_ph_delta(ph.cacl2, soiltype, om, lutum, cp.starch, cp.potato,
  cp.sugarbeet, cp.grass, cp.mais, cp.other)
}
\arguments{
\item{ph.cacl2}{(numeric) The pH-CaCl2 of the soil}

\item{soiltype}{(character) The type of soil}

\item{om}{(numeric) The organic matter content of soil in percentage}

\item{lutum}{(numeric) The percentage lutum present in the soil}

\item{cp.starch}{(numeric) The fraction of starch potatoes in the crop plan}

\item{cp.potato}{(numeric) The fraction of potatoes (excluding starch potatoes) in the crop plan}

\item{cp.sugarbeet}{(numeric) The fraction of sugar beets in the crop plan}

\item{cp.grass}{(numeric) The fracgtion of grass in the crop plan}

\item{cp.mais}{(numeric) The fraction of mais in the crop plan}

\item{cp.other}{(numeric) The fraction of other crops in the crop plan}
}
\description{
This functions calculates the difference between the measured pH and the optimal pH according to the Bemestingsadvies
}
\references{
\href{https://www.handboekbodemenbemesting.nl/nl/handboekbodemenbemesting/Handeling/pH-en-bekalking/Advisering-pH-en-bekalking.htm}{Handboek Bodem en Bemesting tabel 5.1, 5.2 en 5.3}
}
