% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_depth.R
\name{calc_root_depth}
\alias{calc_root_depth}
\title{Determine the root depth of the soil for this crop}
\usage{
calc_root_depth(B_LU_BRP)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}
}
\description{
This function determines the depth of the soil
}
