% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rothc.R
\name{calc_rothc}
\alias{calc_rothc}
\title{RothC model}
\usage{
calc_rothc(
  B_SOILTYPE_AGR,
  A_SOM_LOI,
  A_CLAY_MI,
  A_DEPTH = 0.3,
  event,
  pool_size,
  cor_factors,
  dec_rates = c(10, 0.3, 0.66, 0.02),
  simyears
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_DEPTH}{(numeric) Depth of the soil layer (m)}

\item{event}{(data.table) A data.table with the carbon application events as calculated in calc_events_current or calc_events_minimal}

\item{pool_size}{(numeric) Vector with the initial sizes of the C pools (kg C/ha), order: IOM, CDPM, CRPM, CBIO, CHUM}

\item{cor_factors}{(numeric) A vector with the correction required for RothC, should be in order: temperature, soil moisture, crop cover, grassland renewal}

\item{dec_rates}{rates (numeric) A vector of the decomposition rate constants for the C pools (/year), order: DPM, RPM, BIO and HUM pool; if not provided, default values are used, optional}

\item{simyears}{(numeric) Amount of years for which the simulation should run}
}
\description{
This function contains the Rothemsted carbon model that calculates C decomposition over time
}
\details{
The input parameter event should be a data.table containing the columns time, var, value and method. See ?deSolve::events for the exact format options.
}
\references{
Coleman & Jenkinson (1996) RothC - A model for the turnover of carbon in soil
}
