% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealing.R
\name{calc_sealing}
\alias{calc_sealing}
\title{Calculate soil sealing}
\usage{
calc_sealing(lutum, om)
}
\arguments{
\item{lutum}{(numeric) The percentage lutum present in the soil}

\item{om}{(numeric) The organic matter content of soil in percentage}
}
\description{
This function calculates the risks of soil sealing.  This value can be evaluated by \code{\link{eval_sealing}}
}
