% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{carbon.application.obic}
\alias{carbon.application.obic}
\title{Carbon application table}
\format{
A data.table with 24 rows and 3 columns:
\describe{
  \item{crop_makkink}{Makkink crop category}
  \item{t_manure}{Month in which manure application takes plave, as fraction of a year}
  \item{t_residue}{Month in which the growing period ends and the crop residue enters the soil, as fraction of a year}
}
}
\usage{
carbon.application.obic
}
\description{
This table contains the timing of carbon application via manure and plant residue for the Makkink crop categories
}
\keyword{datasets}
