% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{cf_ind_importance}
\alias{cf_ind_importance}
\title{Helper function to weight and correct the risk and scores}
\usage{
cf_ind_importance(x)
}
\arguments{
\item{x}{The risk or score value to be weighted}
}
\description{
Helper function to weight and correct the risk and scores
}
