% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ov_wenr.R
\name{convert_B_OV_WENR}
\alias{convert_B_OV_WENR}
\title{Convert possible B_OV_WENR values to standardised values}
\usage{
convert_B_OV_WENR(B_OV_WENR)
}
\arguments{
\item{B_OV_WENR}{(numeric or character) Data on soil compaction risk that may have to be converted to string}
}
\description{
This function converts numeric values for OV_WENR to values used by other OBIC functions if numeric values are entered.
}
