% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{crops.makkink}
\alias{crops.makkink}
\title{Makkink correction factor table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 24 rows and 13 columns.
}
\usage{
crops.makkink
}
\description{
This table contains the makkink correction factors for evapo-transpiration per month
}
\details{
\describe{
  \item{crop_makkink}{Makkink crop category}
  \item{1}{Evapotranspiration correction factors for January}
  \item{2}{Evapotranspiration correction factors for February}
  \item{3}{Evapotranspiration correction factors for March}
  \item{4}{Evapotranspiration correction factors for April}
  \item{5}{Evapotranspiration correction factors for May}
  \item{6}{Evapotranspiration correction factors for June}
  \item{7}{Evapotranspiration correction factors for July}
  \item{8}{Evapotranspiration correction factors for August}
  \item{9}{Evapotranspiration correction factors for September}
  \item{10}{Evapotranspiration correction factors for October}
  \item{11}{Evapotranspiration correction factors for November}
  \item{12}{Evapotranspiration correction factors for December}
}
}
\keyword{datasets}
