% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbleability.R
\docType{data}
\name{eval.crumbleability}
\alias{eval.crumbleability}
\title{Coefficient table for evaluating crumbleability}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 16 rows and 4 columns.
}
\usage{
eval.crumbleability
}
\description{
This table contains the coefficients for evaluating the crumbleability. This table is used internally in \code{\link{ind_crumbleability}}
}
\keyword{datasets}
