% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbleability.R
\name{eval_crumbleability}
\alias{eval_crumbleability}
\title{Evaluate the crumbleability}
\usage{
eval_crumbleability(value.crumbleability, crop)
}
\arguments{
\item{value.crumbleability}{(numeric) The value of crumbleability calculated by \code{\link{calc_crumbleability}}}

\item{crop}{(numeric) The crop code (gewascode) from the BRP}
}
\description{
This function evaluates the crumbleability calculated by \code{\link{calc_crumbleability}}
}
