% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen.R
\name{eval_nitrogen}
\alias{eval_nitrogen}
\title{Evaluate nitrogen}
\usage{
eval_nitrogen(value.nlv)
}
\arguments{
\item{value.nlv}{(numeric) The value of NLV  calculated by \code{\link{calc_nlv}}}
}
\description{
This function evaluates the nitrogen content of the soil by using the NLV calculated by \code{\link{calc_nlv}}
}
