% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{eval_ph}
\alias{eval_ph}
\title{Evaluate the difference between de optimum pH and ph of the soil}
\usage{
eval_ph(value.ph.delta)
}
\arguments{
\item{value.ph.delta}{(numeric) The pH difference with the optimal pH.}
}
\description{
This function evaluates the pH of the soil by the difference with the optimum pH. This is calculated in \code{\link{calc_ph_delta}}.
}
