% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphate_availability.R
\name{eval_phosphate_availability}
\alias{eval_phosphate_availability}
\title{Evaluate the phosphate availability}
\usage{
eval_phosphate_availability(value.phosphate.availability)
}
\arguments{
\item{value.phosphate.availability}{(numeric) The value of phosphate availability calculated by \code{\link{calc_phosphate_availability}}}
}
\description{
This function evaluates the phosphate availability calculated by \code{\link{calc_phosphate_availability}}
}
