% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{eval_resistance}
\alias{eval_resistance}
\title{Evaluate the soil resistance}
\usage{
eval_resistance(om)
}
\arguments{
\item{om}{(numeric) The organic matter content of the soil in percentage}
}
\description{
This function evaluates the resistance of the soil against diseases and is inidcatd by the amount of soil life.
}
