% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealing.R
\name{eval_sealing}
\alias{eval_sealing}
\title{Evaluate the soil sealing}
\usage{
eval_sealing(value.sealing, crop)
}
\arguments{
\item{value.sealing}{(numeric) The value of soil sealing calculated by \code{\link{calc_sealing}}}

\item{crop}{(numeric) The crop code (gewascode) from the BRP}
}
\description{
This function evaluates the soil sealing calculated by \code{\link{calc_sealing}}
}
