% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{eval_structure}
\alias{eval_structure}
\title{Evaluate the structure of the soil}
\usage{
eval_structure(eval.crumbleability, eval.sealing)
}
\arguments{
\item{eval.crumbleability}{(numeric) The evaluation value of crumbleability by \code{\link{ind_crumbleability}}}

\item{eval.sealing}{(numeric) the evaluation value of sealing by \code{\link{ind_sealing}}}
}
\description{
This function combines the evalutions for the subject of soil structre and gives an evaluation for it
}
