% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_format_inputs.R
\name{format_aer}
\alias{format_aer}
\title{Convert possible B_AER_CBS values to standardized values}
\usage{
format_aer(B_AER_CBS)
}
\arguments{
\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}
}
\description{
This function formats information of Agricultural Economic Region so it can be understood by other OBIC functions
}
