% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_format_inputs.R
\name{format_gwt}
\alias{format_gwt}
\title{Convert possible B_GWL_CLASS values to standardized values}
\usage{
format_gwt(B_GWL_CLASS, B_AER_CBS)
}
\arguments{
\item{B_GWL_CLASS}{(character) Ground water table classes}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}
}
\value{
A standardized B_GWL_CLASS value as required for the OBIC functions. A character string.
}
\description{
This function assigns a groundwater class if this is unknown (i.e. the value = '-').
If B_AER_CBS is Zuid-Limburg or the corresponding code LG14, '-' becomes 'VIII',
else it becomes 'III'.
}
\examples{
format_gwt(c('sVII', 'sVI', 'IIIb', '-'))

}
