% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosfaatbeschikbaarheid.R
\name{fosfaatbeschikbaarheid}
\alias{fosfaatbeschikbaarheid}
\title{Fosfaatbeschikbaarheid}
\usage{
fosfaatbeschikbaarheid(p_al, p_cacl2, gewas)
}
\arguments{
\item{p_al}{(numeric) Het P-AL gehalte van de bodem}

\item{p_cacl2}{(numeric) Het P-CaCl2 gehalte van de bodem}

\item{gewas}{(character) Het type gewas dat geteelt wordt. Op dit moment is de keuze `gras`, `mais`}
}
\description{
Geeft een waardering voor de fosfaatbeschikbaarheid
}
