% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{ind_aggregatestability}
\alias{ind_aggregatestability}
\title{Calculate the indicator aggregate stability}
\usage{
ind_aggregatestability(D_AS)
}
\arguments{
\item{D_AS}{(numeric) The value of aggregate stability calculated by \code{\link{calc_aggregatestability}}}
}
\description{
This function calculates the indicator for the the aggregate stability of the soil by using the index calculated by \code{\link{calc_aggregatestability}}
}
