% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metals.R
\name{ind_copper}
\alias{ind_copper}
\title{Calculate the indicator for Cu-availability}
\usage{
ind_copper(D_CU)
}
\arguments{
\item{D_CU}{(numeric) The value of Cu-index  calculated by \code{\link{calc_copper_availability}}}
}
\description{
This function calculates the indicator for the the Cu availability in soil by using the Cu-index as calculated by \code{\link{calc_copper_availability}}
}
