% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundwater_recharge.R
\name{ind_gw_target}
\alias{ind_gw_target}
\title{Modify a groundwater recharge indicator with a target}
\usage{
ind_gw_target(I_H_GWR, B_SOILTYPE_AGR, B_GWL_CLASS)
}
\arguments{
\item{I_H_GWR}{(numeric) Evaluation of groundwaterrecharge}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{B_GWL_CLASS}{(character) The groundwater table class}
}
\description{
This function modifiers the groundwater recharge indicator with a correction
factor based on the soiltype and groundwaterclass. Wetter groundwaterclasses
give higher values, peat gives higher values than clay, and clay higher than sand.
}
\examples{
ind_gw_target(0.5, 'zeeklei', 'VIu')


}
