% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen.R
\name{ind_nitrogen}
\alias{ind_nitrogen}
\title{Calculate the indicator for NLV}
\usage{
ind_nitrogen(D_NLV)
}
\arguments{
\item{D_NLV}{(numeric) The value of NLV  calculated by \code{\link{calc_nlv}}}
}
\description{
This function calculates the indicator for the the nitrogen content of the soil by using the NLV calculated by \code{\link{calc_nlv}}
}
