% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nretention.r
\name{ind_nretention_gw}
\alias{ind_nretention_gw}
\title{Calculate the indicator for N retention for groundwater}
\usage{
ind_nretention_gw(D_NUIT_GW)
}
\arguments{
\item{D_NUIT_GW}{(numeric) The value of N leaching calculated by \code{\link{calc_nleach}}}
}
\description{
This function calculates the indicator for the N retention of the soil by using the N leaching to groundwater calculated by \code{\link{calc_nleach}}
}
