% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nretention.r
\name{ind_nretention_ow}
\alias{ind_nretention_ow}
\title{Calculate the indicator for N retention for surface water}
\usage{
ind_nretention_ow(D_NUIT_OW)
}
\arguments{
\item{D_NUIT_OW}{(numeric) The value of N run-off calculated by \code{\link{calc_nleach}}}
}
\description{
This function calculates the indicator for the N retention of the soil by using the N run-off to surface water calculated by \code{\link{calc_nleach}}
}
