% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundwater_recharge.R
\name{ind_permeability}
\alias{ind_permeability}
\title{Calculate the indicator score for the permeability of the top soil}
\usage{
ind_permeability(D_WRI_K)
}
\arguments{
\item{D_WRI_K}{(numeric) The value for top soil permeability (cm/d) as calculated by \code{\link{calc_permeability}}}
}
\description{
This function calculates the indicator score for the permeability of the top soil
}
