% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesticides.R
\name{ind_pesticide_leaching}
\alias{ind_pesticide_leaching}
\title{Calculate an indicator score for pesticide leaching}
\usage{
ind_pesticide_leaching(D_PESTICIDE)
}
\arguments{
\item{D_PESTICIDE}{The fraction of pesticide leached compared to the worst case scenario}
}
\value{
The evaluated score for the soil function to minimize pesticide leaching. A numeric value between 0 and 1.
}
\description{
This function calculates the indicator value for pesticide leaching from a soil
}
\examples{
ind_pesticide_leaching(D_PESTICIDE = 0.7)
ind_pesticide_leaching(D_PESTICIDE  = c(0.4,0.6,0.8,1))
 
}
