% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{ind_resistance}
\alias{ind_resistance}
\title{Calculate indicator for soil resistance}
\usage{
ind_resistance(A_OS_GV)
}
\arguments{
\item{A_OS_GV}{(numeric) The organic matter content of the soil in percentage}
}
\description{
This function calculats the indicator for the resistance of the soil against diseases and is inidcatd by the amount of soil life.
}
