% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winderodibility.R
\name{ind_winderodibility}
\alias{ind_winderodibility}
\title{Calculate indicator for wind erodibility}
\usage{
ind_winderodibility(D_P_DU)
}
\arguments{
\item{D_P_DU}{(numeric) The value for wind erodibility factor (WEF) as calculated by \code{\link{calc_winderodibility}}}
}
\description{
This function calculates the indicator for the resistance of the soil against wind erosion.
}
