% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metals.R
\name{ind_zinc}
\alias{ind_zinc}
\title{Calculate the indicator for Zn-availability}
\usage{
ind_zinc(D_ZN)
}
\arguments{
\item{D_ZN}{(numeric) The value of Zn-index  calculated by \code{\link{calc_zinc_availability}}}
}
\description{
This function calculates the indicator for the the Zn availability in soil by using the Zn-index as calculated by \code{\link{calc_zinc_availability}}
}
