% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{management.obic}
\alias{management.obic}
\title{Relational table linking soil management measures to ecosystem services}
\format{
A data.table with 16 rows and 6 columns:
\describe{
  \item{measure}{The name of measure}
  \item{I_M_SOILFERTILITY}{integrated soil management indicator for soil fertility}
  \item{I_M_CLIMATE}{integrated soil management indicator for soil carbon sequestration}
  \item{I_M_WATERQUALITY}{integrated soil management indicator for water quality}
  \item{I_M_BIODIVERSITY}{Integrated soil management indicator for soil biodiversity}
}
}
\usage{
management.obic
}
\description{
This table assigns which measures positively contribute to the ecosystem services included
}
\keyword{datasets}
