% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{nema.obic}
\alias{nema.obic}
\title{Nematode table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 78 rows and 7 columns.
}
\usage{
nema.obic
}
\description{
This table contains information uses for calculations on nematode species counts
}
\details{
\describe{
  \item{geel}{The intermediate infestation severity count}
  \item{rood}{The count at which a severe infestation is present}
  \item{species}{The species or sometimes genera of the plant parasitic nematode}
  \item{standard}{A boolean indicating whether the species should always be used in calculating the indicator score, regardless of the number of nematodes}
  \item{b}{Growth rate (b) for the evaluate_logistics function}
  \item{v}{v for the evaluate_logistics function, affects the growth rate near the maximum}
}
}
\keyword{datasets}
