% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nretention.R
\docType{data}
\name{nleach_gw_table}
\alias{nleach_gw_table}
\title{Table with fractions of excess N which leaches out to groundwater}
\format{A data.frame with 99 rows and 10 columns:
\describe{
  \item{gewas}{crop type}
  \item{bodem}{soil type}
  \item{ghg}{Lower value for groundwater table (cm-mv)}
  \item{glg}{Upper value for groundwater table (cm-mv)}
  \item{B_GT}{grondwatertrap}
  \item{nf}{N leaching fraction to groundwater (mg NO3/L per kg N overschot/ha/year)}
}}
\usage{
nleach_gw_table
}
\description{
This table contains the fractions of N overshot which leaches out to groundwater, per soil type, crop type, and groundwater table
}
\keyword{datasets}
