% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nretention.R
\docType{data}
\name{nleach_ow_table}
\alias{nleach_ow_table}
\title{Table with fractions of excess N which runs off to surface water}
\format{A data.frame with 99 rows and 10 columns:
\describe{
  \item{gewas}{crop type}
  \item{bodem}{soil type}
  \item{ghg}{Lower value for groundwater table (cm-mv)}
  \item{glg}{Upper value for groundwater table (cm-mv)}
  \item{B_GT}{grondwatertrap}
  \item{nf}{Original values of N run-off fraction to surface water (kg N drain/ha/year per kg N overschot/ha/year)}
}}
\usage{
nleach_ow_table
}
\description{
This table contains the fractions of N overshot which runs off to groundwater, per soil type, crop type, and groundwater table
}
\keyword{datasets}
