% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic.R
\name{obic}
\alias{obic}
\title{Calculate the Open Bodem Index score}
\usage{
obic(dt)
}
\arguments{
\item{dt}{(data.table) A data.table containing the data of the fields to calcualte the OBI}
}
\description{
This functions wraps the functions of the OBIC into one main function to calculate the score for Open Bodem Index (OBI).
}
