% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic.R
\name{obic_field}
\alias{obic_field}
\title{Calculate the Open Bodem Index score for one field}
\usage{
obic_field(
  B_SOILTYPE_AGR,
  B_GWL_CLASS,
  B_SC_WENR,
  B_HELP_WENR,
  B_AER_CBS,
  B_GWL_GLG,
  B_GWL_GHG,
  B_Z_TWO,
  B_LU_BRP,
  A_SOM_LOI,
  A_SAND_MI,
  A_SILT_MI,
  A_CLAY_MI,
  A_PH_CC,
  A_CACO3_IF,
  A_N_RT,
  A_CN_FR,
  A_COM_FR,
  A_S_RT,
  A_N_PMN,
  A_P_AL,
  A_P_CC,
  A_P_WA,
  A_CEC_CO,
  A_CA_CO_PO,
  A_MG_CO_PO,
  A_K_CO_PO,
  A_K_CC,
  A_MG_CC,
  A_MN_CC,
  A_ZN_CC,
  A_CU_CC,
  A_C_BCS = NA,
  A_CC_BCS = NA,
  A_GS_BCS = NA,
  A_P_BCS = NA,
  A_RD_BCS = NA,
  A_EW_BCS = NA,
  A_SS_BCS = NA,
  A_RT_BCS = NA,
  A_SC_BCS = NA,
  M_COMPOST = NA_real_,
  M_GREEN = NA,
  M_NONBARE = NA,
  M_EARLYCROP = NA,
  M_SLEEPHOSE = NA,
  M_DRAIN = NA,
  M_DITCH = NA,
  M_UNDERSEED = NA,
  M_LIME = NA,
  M_NONINVTILL = NA,
  M_SSPM = NA,
  M_SOLIDMANURE = NA,
  M_STRAWRESIDUE = NA,
  M_MECHWEEDS = NA,
  M_PESTICIDES_DST = NA,
  A_RLN_PR_TOT = 0,
  A_RLN_PR_CREN = 0,
  A_RLN_PR_NEG = 0,
  A_RLN_PR_PEN = 0,
  A_RLN_PR_PRA = 0,
  A_RLN_PR_THO = 0,
  A_RLN_PR_FLA = 0,
  A_RLN_PR_FAL = 0,
  A_RLN_PR_PIN = 0,
  A_RLN_PR_PSE = 0,
  A_RLN_PR_VUL = 0,
  A_RLN_PR_DUN = 0,
  A_RLN_PR_ZEA = 0,
  A_RKN_ME_TOT = 0,
  A_RKN_ME_HAP = 0,
  A_RKN_ME_CHIFAL = 0,
  A_RKN_ME_CHI = 0,
  A_RKN_ME_NAA = 0,
  A_RKN_ME_FAL = 0,
  A_RKN_ME_MIN = 0,
  A_RKN_ME_INC = 0,
  A_RKN_ME_JAV = 0,
  A_RKN_ME_ART = 0,
  A_RKN_ME_ARE = 0,
  A_RKN_ME_ARD = 0,
  A_DSN_TR_TOT = 0,
  A_DSN_TR_SIM = 0,
  A_DSN_TR_PRI = 0,
  A_DSN_TR_VIR = 0,
  A_DSN_TR_SPA = 0,
  A_DSN_TR_CYL = 0,
  A_DSN_TR_HOO = 0,
  A_DSN_PA_TER = 0,
  A_DSN_PA_PAC = 0,
  A_DSN_PA_ANE = 0,
  A_DSN_PA_NAN = 0,
  A_DSN_TY_TOT = 0,
  A_DSN_RO_TOT = 0,
  A_DSN_XI_TOT = 0,
  A_DSN_LO_TOT = 0,
  A_DSN_HEM_TOT = 0,
  A_DSN_HEL_TOT = 0,
  A_SN_DI_TOT = 0,
  A_SN_DI_DIP = 0,
  A_SN_DI_DES = 0,
  A_OPN_PA_TOT = 0,
  A_OPN_PA_BUK = 0,
  A_OPN_CY_TOT = 0,
  A_OPN_AP_TOT = 0,
  A_OPN_AP_FRA = 0,
  A_OPN_AP_RIT = 0,
  A_OPN_AP_SUB = 0,
  A_OPN_CR_TOT = 0,
  A_OPN_SU_TOT = 0,
  A_NPN_SA_TOT = 0,
  A_NEMA = NA,
  ID = 1,
  output = "all"
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{B_SC_WENR}{(character) The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006).}

\item{B_HELP_WENR}{(character) The soil type abbreviation, derived from 1:50.000 soil map}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{B_GWL_GLG}{(numeric) The lowest groundwater level averaged over the most dry periods in 8 years in cm below ground level}

\item{B_GWL_GHG}{(numeric) The highest groundwater level averaged over the most wet periods in 8 years in cm below ground level}

\item{B_Z_TWO}{(numeric) The distance between ground level and groundwater level at which the groundwater can supply the soil surface with 2mm water per day (in cm)}

\item{B_LU_BRP}{(numeric) a series with crop codes given the crop rotation plan (source: the BRP)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}

\item{A_CACO3_IF}{(numeric) The carbonate content of the soil (\%)}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{A_CN_FR}{(numeric) The carbon to nitrogen ratio (-)}

\item{A_COM_FR}{(numeric) The carbon fraction of soil organic matter (\%)}

\item{A_S_RT}{(numeric) The total Sulpher content of the soil (in mg S per kg)}

\item{A_N_PMN}{(numeric) The potentially mineralizable N pool (mg N / kg soil)}

\item{A_P_AL}{(numeric) The P-AL content of the soil}

\item{A_P_CC}{(numeric) The plant available P content, extracted with 0.01M CaCl2 (mg / kg)}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water (mg P2O5 / 100 ml soil)}

\item{A_CEC_CO}{(numeric) The cation exchange capacity of the soil (mmol+ / kg), analysed via Cobalt-hexamine extraction}

\item{A_CA_CO_PO}{(numeric) The The occupation of the CEC with Ca (\%)}

\item{A_MG_CO_PO}{(numeric) The The occupation of the CEC with Mg (\%)}

\item{A_K_CO_PO}{(numeric) The occupation of the CEC with K (\%)}

\item{A_K_CC}{(numeric) The plant available K content, extracted with 0.01M CaCl2 (mg / kg)}

\item{A_MG_CC}{(numeric) The plant available Mg content, extracted with 0.01M CaCl2 (ug / kg)}

\item{A_MN_CC}{(numeric) The plant available Mn content, extracted with 0.01M CaCl2 (ug / kg)}

\item{A_ZN_CC}{(numeric) The plant available Zn content, extracted with 0.01M CaCl2 (ug / kg)}

\item{A_CU_CC}{(numeric) The plant available Cu content, extracted with 0.01M CaCl2 (ug / kg)}

\item{A_C_BCS}{(numeric) The presence of visible cracks in the top layer (optional, score 0-1-2)}

\item{A_CC_BCS}{(integer) The crop cover on the surface (optional, score 0-1-2)}

\item{A_GS_BCS}{(numeric) The presence of waterlogged conditions, gley spots (optional, score 0-1-2)}

\item{A_P_BCS}{(numeric) The presence / occurrence of water puddles on the land, ponding (optional, score 0-1-2)}

\item{A_RD_BCS}{(integer) The rooting depth (optional, score 0-1-2)}

\item{A_EW_BCS}{(numeric) The presence of earth worms (optional, score 0-1-2)}

\item{A_SS_BCS}{(integer) The soil structure (optional, score 0-1-2)}

\item{A_RT_BCS}{(numeric) The presence of visible tracks / rutting or trampling on the land (optional, score 0-1-2)}

\item{A_SC_BCS}{(numeric) The presence of compaction of subsoil (optional, score 0-1-2)}

\item{M_COMPOST}{(numeric) The frequency that compost is applied (optional, every x years)}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop (optional, option: yes or no)}

\item{M_NONBARE}{(boolean) A soil measure. Is parcel for 80 percent of the year cultivated and 'green' (optional, option: yes or no)}

\item{M_EARLYCROP}{(boolean) A soil measure. Use of early crop varieties to avoid late harvesting (optional, option: yes or no)}

\item{M_SLEEPHOSE}{(boolean) A soil measure. Is sleephose used for slurry application (optional, option: yes or no)}

\item{M_DRAIN}{(boolean) A soil measure. Are under water drains installed in peaty soils (optional, option: yes or no)}

\item{M_DITCH}{(boolean) A soil measure. Are ditched maintained carefully and slib applied on the land (optional, option: yes or no)}

\item{M_UNDERSEED}{(boolean) A soil measure. Is grass used as second crop in between maize rows (optional, option: yes or no)}

\item{M_LIME}{(boolean) measure. Has field been limed in last three years (option: yes or no)}

\item{M_NONINVTILL}{(boolean) measure. Non inversion tillage (option: yes or no)}

\item{M_SSPM}{(boolean) measure. Soil Structure Protection Measures, such as fixed driving lines, low pressure tires, and light weighted machinery (option: yes or no)}

\item{M_SOLIDMANURE}{(boolean) measure. Use of solid manure (option: yes or no)}

\item{M_STRAWRESIDUE}{(boolean) measure. Application of straw residues (option: yes or no)}

\item{M_MECHWEEDS}{(boolean) measure. Use of mechanical weed protection (option: yes or no)}

\item{M_PESTICIDES_DST}{(boolean) measure. Use of DST for pesticides (option: yes or no)}

\item{A_RLN_PR_TOT}{(numeric) Number of pratylenchus spp. (n / 100g)}

\item{A_RLN_PR_CREN}{(numeric) Number of pratylenchus crenatus (n / 100g)}

\item{A_RLN_PR_NEG}{(numeric) Number of pratylenchus neglectus (n / 100g)}

\item{A_RLN_PR_PEN}{(numeric) Number of pratylenchus penetrans (n / 100g)}

\item{A_RLN_PR_PRA}{(numeric) Number of pratylenchus pratensis (n / 100g)}

\item{A_RLN_PR_THO}{(numeric) Number of pratylenchus thornei (n / 100g)}

\item{A_RLN_PR_FLA}{(numeric) Number of pratylenchus flakkensis (n / 100g)}

\item{A_RLN_PR_FAL}{(numeric) Number of pratylenchus fallax (n / 100g)}

\item{A_RLN_PR_PIN}{(numeric) Number of pratylenchus pinguicaudatus (n / 100g)}

\item{A_RLN_PR_PSE}{(numeric) Number of pratylenchus pseudopratensis (n / 100g)}

\item{A_RLN_PR_VUL}{(numeric) Number of pratylenchus vulnus (n / 100g)}

\item{A_RLN_PR_DUN}{(numeric) Number of pratylenchus dunensis (n / 100g)}

\item{A_RLN_PR_ZEA}{(numeric) Number of pratylenchus zeae (n / 100g)}

\item{A_RKN_ME_TOT}{(numeric) Number of meloidogyne spp. (n / 100g)}

\item{A_RKN_ME_HAP}{(numeric) Number of meloidogyne hapla (n / 100g)}

\item{A_RKN_ME_CHIFAL}{(numeric) Number of meloidogyne chitwoodi/fallax (n / 100g)}

\item{A_RKN_ME_CHI}{(numeric) Number of meloidogyne chitwoodi (n / 100g)}

\item{A_RKN_ME_NAA}{(numeric) Number of meloidogyne naasi (n / 100g)}

\item{A_RKN_ME_FAL}{(numeric) Number of meloidogyne fallax (n / 100g)}

\item{A_RKN_ME_MIN}{(numeric) Number of meloidogyne minor (n / 100g)}

\item{A_RKN_ME_INC}{(numeric) Number of meloidogyne incognita (n / 100g)}

\item{A_RKN_ME_JAV}{(numeric) Number of meloidogyne javanica (n / 100g)}

\item{A_RKN_ME_ART}{(numeric) Number of meloidogyne artiellia (n / 100g)}

\item{A_RKN_ME_ARE}{(numeric) Number of meloidogyne arenaria (n / 100g)}

\item{A_RKN_ME_ARD}{(numeric) Number of meloidogyne ardenensis (n / 100g)}

\item{A_DSN_TR_TOT}{(numeric) Number of trichodoridae spp. (n / 100g)}

\item{A_DSN_TR_SIM}{(numeric) Number of trichodorus similis (n / 100g)}

\item{A_DSN_TR_PRI}{(numeric) Number of trichodorus primitivus (n / 100g)}

\item{A_DSN_TR_VIR}{(numeric) Number of trichodorus viruliferus (n / 100g)}

\item{A_DSN_TR_SPA}{(numeric) Number of trichodorus sparsus (n / 100g)}

\item{A_DSN_TR_CYL}{(numeric) Number of trichodorus cylindricus (n / 100g)}

\item{A_DSN_TR_HOO}{(numeric) Number of trichodorus hooperi (n / 100g)}

\item{A_DSN_PA_TER}{(numeric) Number of paratrichodorus teres (n / 100g)}

\item{A_DSN_PA_PAC}{(numeric) Number of paratrichodorus pachydermus (n / 100g)}

\item{A_DSN_PA_ANE}{(numeric) Number of paratrichodorus anemones (n / 100g)}

\item{A_DSN_PA_NAN}{(numeric) Number of paratrichodorus nanus (n / 100g)}

\item{A_DSN_TY_TOT}{(numeric) Number of tylenchorhynchus spp. (n / 100g)}

\item{A_DSN_RO_TOT}{(numeric) Number of rotylenchus spp. (n / 100g)}

\item{A_DSN_XI_TOT}{(numeric) Number of xiphinema spp. (n / 100g)}

\item{A_DSN_LO_TOT}{(numeric) Number of longidorus spp. (n / 100g)}

\item{A_DSN_HEM_TOT}{(numeric) Number of hemicycliophora spp. (n / 100g)}

\item{A_DSN_HEL_TOT}{(numeric) Number of helicotylenchus spp. (n / 100g)}

\item{A_SN_DI_TOT}{(numeric) Number of ditylenchus spp. (n / 100g)}

\item{A_SN_DI_DIP}{(numeric) Number of ditylenchus dipsaci (n / 100g)}

\item{A_SN_DI_DES}{(numeric) Number of ditylenchus destructor (n / 100g)}

\item{A_OPN_PA_TOT}{(numeric) Number of paratylenchus spp. (n / 100g)}

\item{A_OPN_PA_BUK}{(numeric) Number of paratylenchus bukowinensis (n / 100g)}

\item{A_OPN_CY_TOT}{(numeric) Number of cysteaaltjes (n / 100g)}

\item{A_OPN_AP_TOT}{(numeric) Number of aphelenchoides spp. (n / 100g)}

\item{A_OPN_AP_FRA}{(numeric) Number of aphelenchoides fragariae (n / 100g)}

\item{A_OPN_AP_RIT}{(numeric) Number of aphelenchoides ritzemabosi (n / 100g)}

\item{A_OPN_AP_SUB}{(numeric) Number of aphelenchoides subtenuis (n / 100g)}

\item{A_OPN_CR_TOT}{(numeric) Number of criconematidae spp. (n / 100g)}

\item{A_OPN_SU_TOT}{(numeric) Number of subanguina spp. (n / 100g)}

\item{A_NPN_SA_TOT}{(numeric) Number of saprofage en overige (n / 100g)}

\item{A_NEMA}{(data.table) Long data table with the counted nematodes of a parcel can be used instead of separate nematode entries}

\item{ID}{(character) A field id}

\item{output}{(character) An optional argument to select output: obic_score, scores, indicators, recommendations, or all. (default = all)}
}
\description{
This functions wraps the functions of the OBIC into one main function to calculate the score for Open Bodem Index (OBI) for a single field.
}
