% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic.R
\name{obic_field_dt}
\alias{obic_field_dt}
\title{Calculate the Open Bodem Index score for a data table}
\usage{
obic_field_dt(dt, output = "all")
}
\arguments{
\item{dt}{(data.table) A data.table containing the data of the fields to calculate the OBI}

\item{output}{(character) An optional argument to select output: obic_score, scores, indicators, recommendations, or all. (default = all)}
}
\description{
This functions wraps the functions of the OBIC into one main function to calculate the score for Open Bodem Index (OBI).
In contrast to obic_field, this wrapper can handle a data.table as input.
Multiple sites (distinguished in the column 'ID') can be simulated simultaneously.
}
