% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_new.R
\name{obic_field_test}
\alias{obic_field_test}
\title{Calculate the Open Bodem Index score for one field}
\usage{
obic_field_test(
  B_SOILTYPE_AGR,
  B_GWL_CLASS,
  B_SC_WENR,
  B_HELP_WENR,
  B_AER_CBS,
  B_LU_BRP,
  A_SOM_LOI,
  A_SAND_MI,
  A_SILT_MI,
  A_CLAY_MI,
  A_PH_CC,
  A_CACO3_IF,
  A_N_RT,
  A_CN_FR,
  A_COM_FR,
  A_S_RT,
  A_N_PMN,
  A_P_AL,
  A_P_CC,
  A_P_WA,
  A_CEC_CO,
  A_CA_CO_PO,
  A_MG_CO_PO,
  A_K_CO_PO,
  A_K_CC,
  A_MG_CC,
  A_MN_CC,
  A_ZN_CC,
  A_CU_CC,
  A_C_BCS,
  A_CC_BCS,
  A_GS_BCS,
  A_P_BCS,
  A_RD_BCS,
  A_EW_BCS,
  A_SS_BCS,
  A_RT_BCS,
  A_SC_BCS,
  M_COMPOST,
  M_GREEN,
  M_NONBARE,
  M_EARLYCROP,
  M_SLEEPHOSE,
  M_DRAIN,
  M_DITCH,
  M_UNDERSEED
)
}
\arguments{
\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{B_SC_WENR}{(character) The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006).}

\item{B_HELP_WENR}{(character) The soil type abbreviation, derived from 1:50.000 soil map}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{B_LU_BRP}{(numeric) a series with crop codes given the crop rotation plan (source: the BRP)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_SAND_MI}{(numeric) The}

\item{A_SILT_MI}{(numeric) The}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_PH_CC}{(numeric) The}

\item{A_CACO3_IF}{(numeric) The}

\item{A_N_RT}{(numeric) The}

\item{A_CN_FR}{(numeric) The}

\item{A_COM_FR}{(numeric) The}

\item{A_S_RT}{(numeric) The}

\item{A_N_PMN}{(numeric) The}

\item{A_P_AL}{(numeric) The}

\item{A_P_CC}{(numeric) The}

\item{A_P_WA}{(numeric) The}

\item{A_CEC_CO}{(numeric) The}

\item{A_CA_CO_PO}{(numeric) The}

\item{A_MG_CO_PO}{(numeric) The}

\item{A_K_CO_PO}{(numeric) The}

\item{A_K_CC}{(numeric) The}

\item{A_MG_CC}{(numeric) The}

\item{A_MN_CC}{(numeric) The}

\item{A_ZN_CC}{(numeric) The}

\item{A_CU_CC}{(numeric) The}

\item{A_C_BCS}{(numeric) The presence of visible cracks in the top layer (score 0-1-2)}

\item{A_CC_BCS}{(integer) The crop cover on the surface (score 0-1-2)}

\item{A_GS_BCS}{(numeric) The presence of waterlogged conditions, gley spots (score 0-1-2)}

\item{A_P_BCS}{(numeric) The presence / occurrence of water puddles on the land, ponding (score 0-1-2)}

\item{A_RD_BCS}{(integer) The rooting depth (score 0-1-2)}

\item{A_EW_BCS}{(numeric) The presence of earth worms (score 0-1-2)}

\item{A_SS_BCS}{(integer) The soil structure (score 0-1-2)}

\item{A_RT_BCS}{(numeric) The presence of visible tracks / rutting or trampling on the land (score 0-1-2)}

\item{A_SC_BCS}{(numeric) The presence of compaction of subsoil (score 0-1-2)}

\item{M_COMPOST}{(numeric) The frequency that compost is applied (every x years)}

\item{M_GREEN}{(boolean) measure. are catchcrops sown after main crop (option: yes or no)}

\item{M_NONBARE}{(boolean) measure. is parcel for 80 percent of the year cultivated and 'green' (option: yes or no)}

\item{M_EARLYCROP}{(boolean) measure. use of early crop varieties to avoid late harvesting (option: yes or no)}

\item{M_SLEEPHOSE}{(boolean) measure. is sleepslangbemester used for slurry application (option: yes or no)}

\item{M_DRAIN}{(boolean) measure. are under water drains installed in peaty soils (option: yes or no)}

\item{M_DITCH}{(boolean) measure. are ditched maintained carefully and slib applied on the land (option: yes or no)}

\item{M_UNDERSEED}{(boolean) measure. is grass used as second crop in between maize rows (option: yes or no)}
}
\description{
This functions wraps the functions of the OBIC into one main function to calculate the score for Open Bodem Index (OBI) for a single field.
}
