% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_indicators.R
\name{obic_indicators}
\alias{obic_indicators}
\title{Calculate the indicators for the OBI}
\usage{
obic_indicators(dt.ppr)
}
\arguments{
\item{dt.ppr}{(data.table) The table containg the data needed for OBI}
}
\description{
This wrapper function contains the functions to calculate the indicators of agricultural fields.
}
