% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_preprocessing.R
\name{obic_preprocessing}
\alias{obic_preprocessing}
\title{Preprocess for the OBI}
\usage{
obic_preprocessing(dt)
}
\arguments{
\item{dt}{(data.table) The table containg the data needed for OBI}
}
\description{
This wrapper function contains the functions to preprocess the data of agricultural fields.
}
