% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_recommendations.R
\name{obic_recommendations}
\alias{obic_recommendations}
\title{Recommend measurements for better soil managment}
\usage{
obic_recommendations(dt.score)
}
\arguments{
\item{dt.score}{(data.table)}
}
\description{
This function gives recommendations better soil managament based on the OBI score
}
