% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_recommendations.R
\name{obic_recommendations_bkp}
\alias{obic_recommendations_bkp}
\title{Recommend measurements for better soil management}
\usage{
obic_recommendations_bkp(dt.score, B_LU_BRP, B_SOILTYPE_AGR)
}
\arguments{
\item{dt.score}{(data.table) containing all OBI indicators and scores of a single field}

\item{B_LU_BRP}{(numeric) Cultivation code according to BRP}

\item{B_SOILTYPE_AGR}{(character) Agricultural soil type}
}
\description{
This function returns a list of management recommendations based on OBI scores as part of BodemKwaliteitsPlan.
}
