% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterretention.R
\name{pFpara_class}
\alias{pFpara_class}
\title{Parameter estimation based on class of Staringreeks (Tabel 3, Wosten 2001)}
\usage{
pFpara_class(Pklei, Pleem, Psom, M50)
}
\arguments{
\item{Pklei}{(numeric) The clay (<2um) content of the soil (\%)}

\item{Pleem}{(numeric) The leemt (<50um) content of the soil (\%) Pleem > 0}

\item{Psom}{(numeric) The organic matter content of the soil (\%) Psom > 0}

\item{M50}{(numeric)size of  sand fraction (um)}
}
\description{
Parameter estimation based on class of Staringreeks (Tabel 3, Wosten 2001)
}
