% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obic.R
\name{plot_eval}
\alias{plot_eval}
\title{Visualize the evaluation of a OBI soil function}
\usage{
plot_eval(D_SF, ind_fun)
}
\arguments{
\item{D_SF}{(numeric) The value of a soil index}

\item{ind_fun}{(character) The name of the OBIC function that is visualised}
}
\description{
This function visualizes the evaluation of soil functions as used in the OBIC package.
}
