% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_score.R
\name{score_absolute}
\alias{score_absolute}
\title{Score the indicators for the OBI on an absolute scale}
\usage{
score_absolute(dt.ind)
}
\arguments{
\item{dt.ind}{(data.table) The table containg the data needed for OBI}
}
\description{
This function coalculates the absolute score for the OBI baed on the indicators
}
