% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_score.R
\name{score_relative}
\alias{score_relative}
\title{Score the scoring for the OBI on a relative scale}
\usage{
score_relative(dt.score.abs)
}
\arguments{
\item{dt.score.abs}{(data.table) The table containing the score calculated by \code{\link{score_absolute}}}

\item{id}{(integer) The id of the field to visualize the score}
}
\description{
This function calculates the relative scoring based on the score calculated by \code{\link{score_absolute}}
}
