% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_score.R
\name{score_visualize}
\alias{score_visualize}
\title{Visualize the score}
\usage{
score_visualize(dt.score, id)
}
\arguments{
\item{dt.score}{(data.table) The table containing the score calculated by \code{\link{obic_score}}}

\item{id}{(integer) The id of the field to visualize the score}
}
\description{
Visualize the score of OBIC in absolute terms and relative terms
}
