% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{tbl.ph.delta}
\alias{tbl.ph.delta}
\title{Table with optimal pH for different crop plans}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 136 rows and 10 columns.
}
\usage{
tbl.ph.delta
}
\description{
This table contains the optimal pH for different crop plans and soil types
}
\details{
\describe{
  \item{table}{The original table from Hanboek Bodem en Bemesting}
  \item{lutum.low}{Lower value for A_CLAY_MI}
  \item{lutum.high}{Upper value for A_CLAY_MI}
  \item{om.low}{Lower value for organic matter}
  \item{om.high}{Upper value for organic matter}
  \item{potato.low}{Lower value for fraction potatoes in crop plan}
  \item{potato.high}{Upper value for fraction potatoes in crop plan}
  \item{sugarbeet.low}{Lower value for fraction potatoes in crop plan}
  \item{sugarbeet.high}{Upper value for fraction potatoes in crop plan}
  \item{ph.optimum}{The optimal pH (pH_CaCl2) for this range}   
}

#' @references Handboek Bodem en Bemesting tabel 5.1, 5.2 en 5.3
}
\keyword{datasets}
