% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verkruimelbaarheid.R
\name{verkruimelbaarheid}
\alias{verkruimelbaarheid}
\title{Waardeer de verkruimelbaarheid}
\usage{
verkruimelbaarheid(lutum, leem, os, ph)
}
\arguments{
\item{lutum}{(numeric) Het percentage lutum aanwezig in de bodem}

\item{leem}{(numeric) Het percentage leem aanwezig in de bodem}

\item{os}{(numeric) Het percentage organisch stofgehalte in de bodem}
}
\description{
Met deze functie wordt de verkruimelbaarheid van de bodem gewaardeerd
}
