% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verkruimelbaarheid.R
\name{waardering_lutum}
\alias{waardering_lutum}
\title{Waardering lutum}
\usage{
waardering_lutum(lutum)
}
\arguments{
\item{lutum}{(numeric) Het percentage lutum aanwezig in de bodem}
}
\description{
Waardeer de
}
