% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{weather.obic}
\alias{weather.obic}
\title{Weather table}
\format{
A data.table with 12 rows and 4 columns:
\describe{
  \item{month}{Month of the year}
  \item{A_TEMP_MEAN}{Mean monthly temperature}
  \item{A_PREC_MEAN}{Mean monthly precipitation}
  \item{A_ET_MEAN}{Mean monthly evapo-transpiration}
}
}
\usage{
weather.obic
}
\description{
This table contains the climatic weather data of the Netherlands for the period 1990-2020
}
\keyword{datasets}
