% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_score.R, R/tables.R
\docType{data}
\name{weight_obic}
\alias{weight_obic}
\title{Weight of indicators to calculate integrated scores}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 2 columns.}
\usage{
weight_obic

weight_obic
}
\description{
This table defines the weighting factors (ranging between 0 and 1) of indicator values to calculate integrated scores. 
This table is used internally in \code{\link{obic_score}}

This table defines the weighting factors (ranging between 0 and 1) of indicator values to calculate integrated scores.
}
\keyword{datasets}
