% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{wf}
\alias{wf}
\title{Helper function to weight and correct the risk and scores}
\usage{
wf(x)
}
\arguments{
\item{x}{The risk or score value to be weighted}
}
\description{
Helper function to weight and correct the risk and scores
}
