% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{bouwsteen_tb}
\alias{bouwsteen_tb}
\title{Table with water retention properties of 'bouwstenen'}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 36 rows and 14 columns.
}
\usage{
bouwsteen_tb
}
\description{
This table contains water retention curve parameters and typical mineral composition of 18 'bouwstenen'
}
\details{
\describe{
  \item{bouwsteen}{soil type bouwsteen}
  \item{omschrijving}{description of 'bouwsteen'}
  \item{thres}{residual water content (cm3/cm3). Table 3 of Wosten 2001}
  \item{thsat}{water content at saturation (cm3/cm3). Table 3 of Wosten 2001}
  \item{Ks}{saturated hydraulic conductivity (cm/d). Table 3 of Wosten 2001}
  \item{alpha}{parameter alpha of pF curve (1/cm) Table 3 of Wosten 2001}
  \item{l}{parameter l of pF curve (-). Table 3 of Wosten 2001}
  \item{n}{parameter n of pF curve (-). Table 3 of Wosten 2001}
  \item{sand\%}{sand content (\%) within soil mineral parts. Middle value of Table 1 of Wosten 2001}
  \item{silt\%}{silt content (\%) within soil mineral parts. Middle value of Table 1 of Wosten 2001}
  \item{clay\%}{clay content (\%) within soil mineral parts. Middle value of Table 1 of Wosten 2001}
  \item{OM\%}{organic matter content (\%). Middle value of Table 1 of Wosten 2001}
  \item{bulkdensity}{soil bulk density (g/cm3). Middle value of Table 2 of Wosten 2001}
  \item{M50}{size of sand particles (um). Middle value of Table 2 of Wosten 2001}
}
}
\keyword{datasets}
