% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodemconditiescore.R
\name{calc_bcs}
\alias{calc_bcs}
\title{Calculate the BodemConditieScore}
\usage{
calc_bcs(A_RW_BC, A_BS_BC, A_GV_BC, A_PV_BC, A_AS_BC, A_SV_BC, A_RD_BC,
  A_SS_BC, A_CO_BC, A_OS_GV, D_PH_DELTA, B_LU_BRP, B_BT_AK)
}
\arguments{
\item{A_RW_BC}{(numeric) The presence of earth worms (score 0-1-2)}

\item{A_BS_BC}{(numeric) The presence of compaction of soil structure (score 0-1-2)}

\item{A_GV_BC}{(numeric) The presence of waterlogged conditions (score 0-1-2)}

\item{A_PV_BC}{(numeric) The presence / occurence of water puddles on the land (score 0-1-2)}

\item{A_AS_BC}{(numeric) The presence of visible cracks in the top layer (score 0-1-2)}

\item{A_SV_BC}{(numeric) The presence of visible tracks on the land (score 0-1-2)}

\item{A_RD_BC}{(integer) The rooting depth (score 0-1-2)}

\item{A_SS_BC}{(integer) The soil structure (score 0-1-2)}

\item{A_CO_BC}{(integer) The crop density (score 0-1-2)}

\item{A_OS_GV}{(numeric) The organic matter content of the soil (\%)}

\item{D_PH_DELTA}{(numeric) The pH difference with the optimal pH.}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}

\item{B_BT_AK}{(character) The type of soil}
}
\description{
This function calculates the BodemConditieScore given input from manual observations made in the field.
The individual parameters are scored in three classes: poor (0), neutral (1) or good (2)
More information on this test can be found \href{http://mijnbodemconditie.nl/}{here}
}
\references{
\href{http://mijnbodemconditie.nl/}{mijnbodemconditie.nl}
}
