% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{calc_cec}
\alias{calc_cec}
\title{Calculate a soil fertility index based on the CEC}
\usage{
calc_cec(A_CEC_CO)
}
\arguments{
\item{A_CEC_CO}{(numeric) The cation exchange capacity (mmol+ / kg)}
}
\value{
The capacity of the soil to buffer cations. A numeric value.
}
\description{
This function calculates the capacity of the soil to buffer cations
}
\examples{
calc_cec(A_CEC_CO = 85)
calc_cec(A_CEC_CO = c(85,125,326))

}
