% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{calc_cec}
\alias{calc_cec}
\title{Calculate the index for the CEC-occupation}
\usage{
calc_cec(A_CEC_CO, A_K_CEC, A_CA_CEC, A_MG_CEC, advice)
}
\arguments{
\item{A_CEC_CO}{(numeric) The cation exchange capacity (mmol+ / kg)}

\item{A_K_CEC}{(numeric) The occupation of the CEC with K (\%)}

\item{A_CA_CEC}{(numeric) The occupation of the CEC with Mg (\%)}

\item{A_MG_CEC}{(numeric) The occupation of the CEC with Ca (\%)}

\item{advice}{(character) Optional parameter to select CEC index for soil structure or fertility. Options: fertility_index or structure_index}
}
\description{
This function calculates the CEC-buffer index for a soil
}
