% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metals.R
\name{calc_copper_availability}
\alias{calc_copper_availability}
\title{Calculate the availability of the metal Cu}
\usage{
calc_copper_availability(
  B_LU_BRP,
  A_SOM_LOI,
  A_CLAY_MI,
  A_K_CC,
  A_MN_CC,
  A_CU_CC
)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_K_CC}{(numeric) The plant available potassium, extracted with 0.01M CaCl2 (mg / kg),}

\item{A_MN_CC}{(numeric) The plant available Mn content, extracted with 0.01M CaCl2 (ug / kg)}

\item{A_CU_CC}{(numeric) The plant available Cu content, extracted with 0.01M CaCl2 (ug / kg)}
}
\description{
This function calculates the availability of Cu for plant uptake
}
