% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metals.R
\name{calc_copper_availability}
\alias{calc_copper_availability}
\title{Calculate the availability of the metal Cu}
\usage{
calc_copper_availability(A_CU_CC, B_LU_BRP, B_BT_AK)
}
\arguments{
\item{A_CU_CC}{The plant available Cu content, extracted with 0.01M CaCl2 (mg / kg)}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}

\item{B_BT_AK}{(character) The type of soil}
}
\description{
This function calculates the availability of Cu for plant uptake
}
