% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_model.R
\name{calc_crop_rotation}
\alias{calc_crop_rotation}
\title{Determine the crop rotation of a field}
\usage{
calc_crop_rotation(B_LU_BRP, M_GREEN = FALSE, effectivity = TRUE)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop, optional}

\item{effectivity}{(boolean) A vector that tells whether the catch crop was effective (i.e. did it grow sufficiently), optional}
}
\description{
This function determines crop cover and Makkink correction factors based on the cultivated crops
}
