% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbleability.R
\name{calc_crumbleability}
\alias{calc_crumbleability}
\title{Calculate the crumbleability}
\usage{
calc_crumbleability(lutum, om, ph)
}
\arguments{
\item{lutum}{(numeric) The percentage lutum available of the soil}

\item{om}{(numeric) The organic matter content of soil in percentage}

\item{ph}{(numeric) The pH of the soil}
}
\description{
This function calculates the crumbleability. This value can be evaluated by \code{\link{eval_crumbleability}}
}
