% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass_age.R
\name{calc_grass_age}
\alias{calc_grass_age}
\title{Calculate the average age of the grass}
\usage{
calc_grass_age(ID, B_LU_BRP)
}
\arguments{
\item{ID}{(numeric) The ID of the field}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}
}
\description{
This function calculates the average age of the grass
}
