% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makkink.R
\name{calc_makkink}
\alias{calc_makkink}
\title{Add Makkink correction factors and crop cover to crop rotation table}
\usage{
calc_makkink(B_LU_BRP)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}
}
\value{
A datatable with the crop dependent Makkink correction factor per month. 
Output is a single data.table with for each B_LU_BRP code the monthly correction factor.
Columns of the data.table are: crop_makkink, month, year, mcf and crop_cover.
}
\description{
This function adds Makkink correction factors for ET and crop cover to the crop rotation table
}
\examples{
calc_makkink(B_LU_BRP = 265)
calc_makkink(B_LU_BRP = c(265,1019))

}
