% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makkink.R
\name{calc_makkink}
\alias{calc_makkink}
\title{Add Makkink correction factors and crop cover to crop rotation table}
\usage{
calc_makkink(B_LU_BRP)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}
}
\description{
This function adds Makkink correction factors for ET and crop cover to the crop rotation table
}
