% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{calc_management}
\alias{calc_management}
\title{Calculate the 'performance' of sustainable soil management}
\usage{
calc_management(A_OS_GV, B_LU_BRP, B_BT_AK, B_GT, D_OS_BAL, D_CP_GRASS,
  D_CP_POTATO, D_CP_RUST, D_CP_RUSTDEEP, D_GA, M_M4, M_M6, M_M10, M_M11,
  M_M12, M_M13, M_M14, M_M15)
}
\arguments{
\item{A_OS_GV}{(numeric) The organic matter content of the soil (in procent)}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}

\item{B_BT_AK}{(character) The type of soil}

\item{B_GT}{(character) The groundwater table class}

\item{D_OS_BAL}{(numeric) The organic matter balance of the soil (in kg EOS per ha)}

\item{D_CP_GRASS}{(numeric) The fraction grassland in crop rotation}

\item{D_CP_POTATO}{(numeric) The fraction potato crops in crop rotation}

\item{D_CP_RUST}{(numeric) The fraction rustgewassen in crop rotation}

\item{D_CP_RUSTDEEP}{(numeric) The fraction diepe rustgewassen in crop rotation (-)}

\item{D_GA}{(numeric) The age of the grassland (years)}

\item{M_M4}{(boolean) measure 4. is the soil ploughed (option: yes or no)}

\item{M_M6}{(boolean) measure 6. are catchcrops sown after main crop (option: yes or no)}

\item{M_M10}{(boolean) measure 10. is parcel for 80 percent of the year cultivated and 'green' (option: yes or no)}

\item{M_M11}{(boolean) measure 11. use of early crop varieties to avoid late harvesting (option: yes or no)}

\item{M_M12}{(boolean) measure 12. is sleepslangbemester used for slurry application (option: yes or no)}

\item{M_M13}{(boolean) measure 13. are under water drains installed in peaty soils (option: yes or no)}

\item{M_M14}{(boolean) measure 14. are ditched maintained carefully and slib applied on the land (option: yes or no)}

\item{M_M15}{(boolean) measure 15. is grass used as second crop in between maize rows (option: yes or no)}
}
\description{
This function evaluates the contribution of sustainable soil management following the Label Sustainable Soil Management.
}
