% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen.R
\name{calc_nlv}
\alias{calc_nlv}
\title{Calculate the NLV}
\usage{
calc_nlv(A_N_TOT, D_OC, B_LU_BRP, B_BT_AK, D_BDS, A_CN_RAT, D_GA)
}
\arguments{
\item{A_N_TOT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{D_OC}{(numeric) The organic carbon content of the soil in kg C / ha}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}

\item{B_BT_AK}{(character) The type of soil}

\item{D_BDS}{(numeric) The bulk density of the soil in kg / m3}

\item{A_CN_RAT}{(numeric)}

\item{D_GA}{(numeric) The age of the grass if present}
}
\description{
This function calculates the NLV (nitrogen producing capacity) for the soil
}
