% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen.R
\name{calc_nlv}
\alias{calc_nlv}
\title{Calculate the NLV}
\usage{
calc_nlv(n.org, c.org, crop, soiltype, bulk_density, cn_ratio, grass.age)
}
\arguments{
\item{n.org}{(numeric) The organic nitrogen content of the soil in g  N / kg}

\item{c.org}{(numeric) The organic carbon content of the soil in kg C / ha}

\item{crop}{(numeric) The crop code (gewascode) from the BRP}

\item{soiltype}{(character) The type of soil}

\item{bulk_density}{(numeric) The bulk density of the soil in kg / m3}

\item{cn_ratio}{(numeric)}

\item{grass.age}{(numeric) The age of the grass if present}
}
\description{
This function calculates the NLV (nitrogen producing capacity) for the soil
}
