% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{calc_ph_delta}
\alias{calc_ph_delta}
\title{Calculate the difference between pH and optimum}
\usage{
calc_ph_delta(
  A_PH_CC,
  B_BT_AK,
  A_CLAY_MI,
  A_OS_GV,
  D_CP_STARCH,
  D_CP_POTATO,
  D_CP_SUGARBEET,
  D_CP_GRASS,
  D_CP_MAIS,
  D_CP_OTHER,
  B_LU_BRP
)
}
\arguments{
\item{A_PH_CC}{(numeric) The pH-CaCl2 of the soil}

\item{B_BT_AK}{(character) The type of soil}

\item{A_CLAY_MI}{(numeric) The percentage A_CLAY_MI present in the soil}

\item{A_OS_GV}{(numeric) The organic matter content of soil in percentage}

\item{D_CP_STARCH}{(numeric) The fraction of starch potatoes in the crop plan}

\item{D_CP_POTATO}{(numeric) The fraction of potatoes (excluding starch potatoes) in the crop plan}

\item{D_CP_SUGARBEET}{(numeric) The fraction of sugar beets in the crop plan}

\item{D_CP_GRASS}{(numeric) The fracgtion of grass in the crop plan}

\item{D_CP_MAIS}{(numeric) The fraction of mais in the crop plan}

\item{D_CP_OTHER}{(numeric) The fraction of other crops in the crop plan}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}
}
\description{
This functions calculates the difference between the measured pH and the optimal pH according to the Bemestingsadvies
}
\references{
\href{https://www.handboekbodemenbemesting.nl/nl/handboekbodemenbemesting/Handeling/pH-en-bekalking/Advisering-pH-en-bekalking.htm}{Handboek Bodem en Bemesting tabel 5.1, 5.2 en 5.3}
}
