% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphate_availability.R
\name{calc_phosphate_availability}
\alias{calc_phosphate_availability}
\title{Calculate the phosphate availability (PBI)}
\usage{
calc_phosphate_availability(p_al, p_cacl2, crop)
}
\arguments{
\item{p_al}{(numeric) The P-AL content of the soil}

\item{p_cacl2}{(numeric) The P-CaCl2 content of the soil}

\item{crop}{(numeric) The crop code (gewascode) from the BRP}
}
\description{
This function calculates the phosphate availability. This value can be evaluated by \code{\link{eval_phosphate_availability}}
}
