% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphate_availability.R
\name{calc_phosphate_availability}
\alias{calc_phosphate_availability}
\title{Calculate the phosphate availability (PBI)}
\usage{
calc_phosphate_availability(
  A_P_AL = NULL,
  A_P_CC = NULL,
  A_P_WA = NULL,
  B_LU_BRP
)
}
\arguments{
\item{A_P_AL}{(numeric) The P-AL content of the soil}

\item{A_P_CC}{(numeric) The P-CaCl2 content of the soil}

\item{A_P_WA}{(numeric) The P-content of the soil extracted with water}

\item{B_LU_BRP}{(numeric) The crop code from the BRP}
}
\description{
This function calculates the phosphate availability. This value can be evaluated by \code{\link{ind_phosphate_availability}}
}
