% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation_surplus.R
\name{calc_psp}
\alias{calc_psp}
\title{Calculate the precipitation surplus}
\usage{
calc_psp(B_LU_BRP, M_GREEN)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{M_GREEN}{(boolean) A soil measure. Are catch crops sown after main crop (optional, option: yes or no)}
}
\description{
This function calculates the precipitation surplus (in mm / ha) given the crop rotation plan.
}
