% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_rotation.R
\name{calc_rotation_fraction}
\alias{calc_rotation_fraction}
\title{Calculates the fraction in the crop rotation}
\usage{
calc_rotation_fraction(ID, B_LU_BRP, crop)
}
\arguments{
\item{ID}{(numeric) The ID of the field}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}

\item{crop}{(character) The crop to check for}
}
\description{
This function calculates the fraction present in the crop rotation
}
