% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sulpher.R
\name{calc_sbal_arable}
\alias{calc_sbal_arable}
\title{Calculate the indicator for delta S-balans arable}
\usage{
calc_sbal_arable(D_SLV, B_LU_BRP, B_SOILTYPE_AGR, B_AER_CBS)
}
\arguments{
\item{D_SLV}{(numeric) The value of SLV  calculated by \code{\link{calc_slv}}}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}

\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}
}
\description{
This function calculates the change in S-balans compared to averaged S-supply as given in fertilizer recommendation systems.
}
