% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealing.R
\name{calc_sealing_risk}
\alias{calc_sealing_risk}
\title{Calculate soil sealing risk}
\usage{
calc_sealing_risk(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The organic matter content of soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}
}
\description{
This function calculates the risks of soil sealing.  This value can be evaluated by \code{\link{ind_sealing}}
}
