% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealing.R
\name{calc_sealing_risk}
\alias{calc_sealing_risk}
\title{Calculate soil sealing risk}
\usage{
calc_sealing_risk(A_CLAY_MI, A_OS_GV)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The percentage A_CLAY_MI present in the soil}

\item{A_OS_GV}{(numeric) The organic matter content of soil in percentage}
}
\description{
This function calculates the risks of soil sealing.  This value can be evaluated by \code{\link{ind_sealing}}
}
