% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sulpher.R
\name{calc_slv}
\alias{calc_slv}
\title{Calculate the SLV}
\usage{
calc_slv(A_OS_GV, A_S_TOT, B_LU_BRP, B_BT_AK, B_LG_CBS, D_BDS)
}
\arguments{
\item{A_OS_GV}{(numeric) The organic matter content of the soil (in procent)}

\item{A_S_TOT}{(numeric) The total Sulpher content of the soil (in mg S per kg)}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}

\item{B_BT_AK}{(character) The type of soil}

\item{B_LG_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{D_BDS}{(numeric) The bulk density of the soil (in kg per m3)}
}
\description{
This function calculates a S-balance given the SLV (Sulpher supplying capacity) of a soil
}
