% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sulpher.R
\name{calc_slv}
\alias{calc_slv}
\title{Calculate the SLV}
\usage{
calc_slv(B_LU_BRP, B_SOILTYPE_AGR, B_AER_CBS, A_SOM_LOI, A_S_RT, D_BDS)
}
\arguments{
\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{B_SOILTYPE_AGR}{(character) The type of soil}

\item{B_AER_CBS}{(character) The agricultural economic region in the Netherlands (CBS, 2016)}

\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (in procent)}

\item{A_S_RT}{(numeric) The total Sulpher content of the soil (in mg S per kg)}

\item{D_BDS}{(numeric) The bulk density of the soil (in kg per m3)}
}
\description{
This function calculates a S-balance given the SLV (Sulpher supplying capacity) of a soil
}
