% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterretention.R
\name{calc_waterretention}
\alias{calc_waterretention}
\title{Calculate indicators for water retention in topsoil}
\usage{
calc_waterretention(A_CLAY_MI, A_SAND_MI, A_SILT_MI, A_OS_GV,
  type = "plant available water")
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{A_OS_GV}{(numeric) The organic matter content of the soil (\%)}

\item{type}{(character) The type of waterretention index. Options include c('wilting point','field capacity','water holding capacity','plant available water','Ksat')}
}
\description{
This function calculates different kind of Water Retention Indices given the continuous pedotransferfunctions of Wosten et al. (2001)
These include : 'wilting point','field capacity','water holding capacity','plant available water' and 'Ksat'
}
