% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metals.R
\name{calc_zinc_availability}
\alias{calc_zinc_availability}
\title{Calculate the availability of the metal Zinc}
\usage{
calc_zinc_availability(A_ZN_CC, B_LU_BRP, B_BT_AK, A_PH_CC)
}
\arguments{
\item{A_ZN_CC}{The plant available Zn content, extracted with 0.01M CaCl2 (mg / kg)}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}

\item{B_BT_AK}{(character) The type of soil}

\item{A_PH_CC}{(numeric) The acidity of the soil, determined in 0.01M CaCl2 (-)}
}
\description{
This function calculates the availability of Zn for plant uptake
}
