% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{column_description.obic}
\alias{column_description.obic}
\title{Column description for the OBIC}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 222 rows and 6 columns.
}
\usage{
column_description.obic
}
\description{
This table defines the columns used in the OBIC and which unit is used
}
\details{
\describe{
  \item{column}{The column name used in OBIC}
  \item{type}{The type of column}
  \item{description_nl}{A description of the column in Dutch}
  \item{description_en}{A description of the column in English}
  \item{unit}{The unit used for this column}
  \item{method}{The method to measure/obtain the values for this column}
}
}
\keyword{datasets}
