% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{crops.obic}
\alias{crops.obic}
\title{Linking table between crops and different functions in OBIC}
\format{A data.frame with 371 rows and 5 columns:
\describe{
  \item{crop_name}{The name of the crop}
  \item{crop_code}{The BRP gewascode of the crop}
  \item{crop_crumbleability}{The category for this crop at crumbleablity}
  \item{crop_phosphate}{The category for this crop at phosphate availability}
  \item{crop_sealing}{The category for this crop at soil sealing}
  \item{crop_n}{The category for this crop at nitrogen}
}}
\usage{
crops.obic
}
\description{
This table helps to link the different crops in the OBIC functions with the crops selected by the user
}
\keyword{datasets}
