% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops_obic.R
\docType{data}
\name{crops.obic}
\alias{crops.obic}
\title{Linking table between crops and different functions in OBIC}
\format{A data.frame with 371 rows and 5 columns:
\describe{
  \item{crop_name}{The name of the crop}
  \item{crop_code}{The BRP gewascode of the crop}
  \item{crumbleability}{The category for this crop in crumbleablity}
}}
\usage{
crops.obic
}
\description{
This table helps to link the different crops in the OBIC functions with the crops selected by the user
}
\keyword{datasets}
