% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{crops.obic}
\alias{crops.obic}
\title{alias of crops_obic}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 460 rows and 22 columns.
}
\usage{
crops.obic
}
\description{
alias of crops_obic
}
\keyword{datasets}
