% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_logistic}
\alias{evaluate_logistic}
\title{Evaluate using the general logistice function}
\usage{
evaluate_logistic(x, b, x0, v)
}
\arguments{
\item{x}{(numeric) The values of a calc function to be converted to an evaluation}

\item{b}{(numeric) The growth rate}

\item{x0}{(numeric) The offset of the x-axis}

\item{v}{(numeric) Affects the growth rate near the maximum}
}
\description{
This function evaluates the calculated values from an indicator using a general logistic function
}
\references{
\url{https://en.wikipedia.org/wiki/Generalised_logistic_function}
}
