% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gwt.R
\name{format_gwt}
\alias{format_gwt}
\title{Convert possible B_GT values to standardised values}
\usage{
format_gwt(B_GT)
}
\arguments{
\item{B_GT}{(sting) Ground water table classes}
}
\description{
This function formats ground water table information so it can be understood by other OBIC functions
}
