% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_format_inputs.R
\name{format_gwt}
\alias{format_gwt}
\title{Convert possible B_GT values to standardized values}
\usage{
format_gwt(B_GWL_CLASS)
}
\arguments{
\item{B_GWL_CLASS}{(character) Ground water table classes}
}
\description{
This function formats ground water table information so it can be understood by other OBIC functions
}
