% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_format_inputs.R
\name{format_gwt}
\alias{format_gwt}
\title{Convert possible B_GWL_CLASS values to standardized values}
\usage{
format_gwt(B_GWL_CLASS)
}
\arguments{
\item{B_GWL_CLASS}{(character) Ground water table classes}
}
\value{
A standardized B_GWL_CLASS value as required for the OBIC functions. A character string.
}
\description{
This function formats ground water table information so it can be understood by other OBIC functions
}
\examples{
format_gwt(c('sVII', 'sVI'))
format_gwt(c('sVII', 'sVI','GtII', 'GtI'))

}
