% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compaction.R
\name{ind_compaction}
\alias{ind_compaction}
\title{Calculate indicator for subsoil compaction}
\usage{
ind_compaction(B_OV_WENR)
}
\arguments{
\item{B_OV_WENR}{(numeric) The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006)}
}
\description{
This function calculats the indicator for the risk for soil compaction of the subsoil.
derived from van den Akker et al. (2013) Risico op ondergrondverdichting in het landelijk gebied in kaart, 
Alterra-rapport 2409, Alterra, Wageningen University and Research Centre,
}
