% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundwater_recharge.R
\name{ind_gw_recharge}
\alias{ind_gw_recharge}
\title{Calculate the indicator for the groundwater recharge through precipitation surplus in the Netherlands}
\usage{
ind_gw_recharge(D_PSP)
}
\arguments{
\item{D_PSP}{(numeric) The precipitation surplus per crop  calculated by \code{\link{calc_gw_recharge}}}
}
\description{
This function calculates the indicator for the the groundwater recharge through precipitation surplus calculated by \code{\link{calc_gw_recharge}}
}
