% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen_efficiency.R
\name{ind_n_efficiency}
\alias{ind_n_efficiency}
\title{Calculate an indicator value for nitrogen use efficiency and leaching based on N surplus}
\usage{
ind_n_efficiency(D_NLEACH)
}
\arguments{
\item{D_NLEACH}{(numeric) The value of N leaching calculated by \code{\link{calc_n_efficiency}}}
}
\value{
The evaluated score for the soil function to enhance the nitrogen use efficiency. A numeric value between 0 and 1.
}
\description{
This function gives an indicator value for nitrogen use efficiency calculated by \code{\link{calc_n_efficiency}}, this function makes use of \code{\link{ind_nretention}}
}
\examples{
ind_n_efficiency(D_NLEACH = 50)
ind_n_efficiency(D_NLEACH = c(5,15,25,75))
 
}
