% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesticides.R
\name{ind_pesticide_leaching}
\alias{ind_pesticide_leaching}
\title{Calculate an indicator score for pesticide leaching}
\usage{
ind_pesticide_leaching(D_PESTICIDE, B_GWL_CLASS)
}
\arguments{
\item{D_PESTICIDE}{The fraction of pesticide leached compared to the worst case scenario}

\item{B_GWL_CLASS}{(character) The groundwater table class}
}
\description{
This function calculates the indicator value for pesticide leaching from a soil
}
