% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{ind_ph}
\alias{ind_ph}
\title{Calculate the indicator for pH}
\usage{
ind_ph(D_PH_DELTA)
}
\arguments{
\item{D_PH_DELTA}{(numeric) The pH difference with the optimal pH.}
}
\description{
This function calculates the indicator for the pH of the soil by the difference with the optimum pH. This is calculated in \code{\link{calc_ph_delta}}.
}
