% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{ind_pmn}
\alias{ind_pmn}
\title{Calculate the indicator for microbial biological activity}
\usage{
ind_pmn(D_PMN)
}
\arguments{
\item{D_PMN}{(numeric) The value of PMN  calculated by \code{\link{calc_pmn}}}
}
\description{
This function calculates the indicator that assess the microbial biological activity of the soil by using the PMN calculated by \code{\link{calc_pmn}}
}
