% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation_surplus.R
\name{ind_psp}
\alias{ind_psp}
\title{Calculate indicator for precipitation surplus}
\usage{
ind_psp(D_PSP)
}
\arguments{
\item{D_PSP}{(numeric) The precipitation surplus per crop  calculated by \code{\link{calc_psp}}}
}
\description{
This function calculates the indicator value for precipitation surplus
}
