% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealing.R
\name{ind_sealing}
\alias{ind_sealing}
\title{Calculate the soil sealing indicator}
\usage{
ind_sealing(D_SE, B_LU_BRP)
}
\arguments{
\item{D_SE}{(numeric) The value of soil sealing calculated by \code{\link{calc_sealing_risk}}}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}
}
\description{
This function calculates the indicator for the soil sealing calculated by \code{\link{calc_sealing_risk}}
}
