% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workability.R
\name{ind_workability}
\alias{ind_workability}
\title{Calculate indicator for workability}
\usage{
ind_workability(D_P_WO, B_LU_BRP)
}
\arguments{
\item{D_P_WO}{(numeric) The value of the relative (workable) season length calculated by \code{\link{calc_workability}}}

\item{B_LU_BRP}{(numeric) The crop code (gewascode) from the BRP}
}
\description{
This function calculates the indicator for the workability of the soil expressed as the period in which the soil can be worked without
inflicting structural damage that cannot be restored by the regular management on the farm.
}
