% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_recommendations.R, R/tables.R
\docType{data}
\name{maatregel.obic}
\alias{maatregel.obic}
\title{applicability range of measures including literature based estimates of effect on soil indicators}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 252 rows and 18 columns.}
\usage{
maatregel.obic

maatregel.obic
}
\description{
This table defines the effects of 11 measures on soil indicators. 
This table is used internally in \code{\link{obic_evalmeasure}}

This table defines the effects of 11 measures on soil indicators
}
\keyword{datasets}
