% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nretention.R
\docType{data}
\name{nleach_ow_table}
\alias{nleach_ow_table}
\title{Table with fractions of excess N which runs off to surface water}
\format{A data.frame with 99 rows and 10 columns:
\describe{
  \item{gt}{grondwatertrap}
  \item{gewas}{crop type}
  \item{bodem}{soil type}
  \item{nloss}{Original values of N run-off fraction to surface water (kg N drain/ha/year per kg N overschot/ha/yaer)}
  \item{ghg}{Lower value for groundwater table}
  \item{glg}{Upper value for groundwater table}
  \item{nloss2}{Original values + manually assigned values (for GT1) of N leaching fraction to groundwater}
  \item{pred}{Predicted value of N leaching fraction to groundwater}
  \item{filled}{Original values + predicted values (for NA cells) of N leaching fraction to groundwater}
  \item{orifill}{whether the value is the original ("model"), filled ("gevuld"), or manually assigned ("voorgeschreven"))}
}}
\usage{
nleach_ow_table
}
\description{
This table contains the fractions of N overshot which runs off to groundwater, per soil type, crop type, and groundwater table
}
\keyword{datasets}
