% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic.R
\name{obic}
\alias{obic}
\title{Calculate the Open Bodem Index score}
\usage{
obic(dt, add_relative_score = TRUE, add_recommendations = TRUE)
}
\arguments{
\item{dt}{(data.table) A data.table containing the data of the fields to calcualte the OBI}

\item{add_relative_score}{(logical) Should the relative score be calculated? Defaults to TRUE}

\item{add_recommendations}{(logical) Should recommendations be given to improve the OBI score? Defaults to TRUE}
}
\description{
This functions wraps the functions of the OBIC into one main function to calculate the score for Open Bodem Index (OBI).
}
\details{
The argument add_relative_score calculates relative perforamce of the agricultural fields in the OBI. 
To have a meaningfull relative score a large number of fields need be given as input in dt.
}
