% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_recommendations.R
\name{obic_evalmeasure}
\alias{obic_evalmeasure}
\title{Evaluate effects of measures}
\usage{
obic_evalmeasure(dt.score)
}
\arguments{
\item{dt.score}{(data.table)}
}
\description{
This function quantifies the effects of 11 soil measures on the OBI score
}
