% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_new.R
\name{obic_field}
\alias{obic_field}
\title{Calculate the Open Bodem Index score for one field}
\usage{
obic_field(
  refid,
  B_LU_BRP,
  B_SC_WENR,
  B_GWL_CLASS,
  B_SOILTYPE_AGR,
  A_SOM_LOI,
  A_CLAY_MI,
  A_PH_CC
)
}
\arguments{
\item{refid}{(character) a field id}

\item{B_LU_BRP}{(numeric) The crop code from the BRP}

\item{B_SC_WENR}{(character) The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006).}

\item{B_GWL_CLASS}{(character) The groundwater table class}

\item{B_SOILTYPE_AGR}{(character) The agricultural type of soil}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%)}

\item{A_PH_CC}{(numeric) The acidity of the soil, determined in 0.01M CaCl2 (-)}
}
\description{
This functions wraps the functions of the OBIC into one main function to calculate the score for Open Bodem Index (OBI) for a single field.
}
