% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_indicators.R
\name{obic_indicators}
\alias{obic_indicators}
\title{Calculate the indicators for the OBI}
\usage{
obic_indicators(dt.ppr, dt_nema = NA)
}
\arguments{
\item{dt.ppr}{(data.table) The table containing the data needed for OBI}

\item{dt_nema}{(data.table) The table containing the counted plant parasitic nematodes}
}
\description{
This wrapper function contains the functions to calculate the indicators of agricultural fields.
}
