% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_recommendations.R
\name{obic_recommendations}
\alias{obic_recommendations}
\title{Recommend measurements for better soil managment}
\usage{
obic_recommendations(dt.recom, extensive = FALSE)
}
\arguments{
\item{dt.recom}{(data.table)}

\item{extensive}{(boolean) whether the output table includes evaluation scores of each  measures (TRUE) or only names of top 3 measures}
}
\description{
This function gives recommendations better soil management based on the OBI score
}
