% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_recommendations.R
\name{obic_recommendations}
\alias{obic_recommendations}
\title{Recommend measurements for better soil managment}
\usage{
obic_recommendations(dt.recom)
}
\arguments{
\item{dt.recom}{(data.table) The results from \code{\link{obic_evalmeasure}}}
}
\description{
This function gives recommendations better soil management based on the OBI score
}
