% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obic_score.R
\name{obic_score}
\alias{obic_score}
\title{Score the indicators for the OBI}
\usage{
obic_score(dt.ind)
}
\arguments{
\item{dt.ind}{(data.table) The table containg the data needed for OBI}
}
\description{
This wrapper function contains the functions to perform the scoring
}
