% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterretention.R
\name{pF_curve}
\alias{pF_curve}
\title{Water retention curve}
\usage{
pF_curve(head, thetaR, thetaS, alfa, n)
}
\arguments{
\item{head}{(numeric)  suction pressure ([L] or cm of water)}

\item{thetaR}{(numeric) residual water content (cm3/cm3)}

\item{thetaS}{(numeric) saturated water content (cm3/cm3)}

\item{alfa}{(numeric)  related to the inverse of the air entry suction, alfa > 0 (1/cm)}

\item{n}{(numeric)  a measure of the pore-size distribution, n>1, dimensionless}
}
\value{
theta (numeric) water content (cm3/cm3)
}
\description{
This function compute water content at given pressure head, using Van Genuchten water retention curve
}
