% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{waterstress.obic}
\alias{waterstress.obic}
\title{Linking table between crops, soils, groundwatertables and water induced stresses in OBIC}
\format{
A data.frame with x rows and y columns:
\describe{
  \item{cropname}{The name of the crop}
  \item{soilunit}{The category for this soil, derived from 1:50.000 soil map}
  \item{gt}{The class describing mean highest and lowest groundwater table, derived from 1:50.000 soil map}
  \item{droughtstress}{The mean yield reduction due to drought (in percentage)}
  \item{wetnessstress}{The mean yield reduction due to watersurplus (in percentage)}
  \item{waterstress}{The mean combined effect water stress (due to deficiency or excess of water)}
}
}
\usage{
waterstress.obic
}
\description{
This table helps to link the different crops in the OBIC functions with the crops selected by the user
}
\keyword{datasets}
