% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{weight.obic}
\alias{weight.obic}
\title{Weight of indicators to calculate integrated scores}
\format{
A data.frame with 20 rows and 2 columns:
\describe{
  \item{var}{The name of the weight}
  \item{weight}{weighing factor}
}
}
\usage{
weight.obic
}
\description{
This table defines the weighting factors (ranging between 0 and 1) of indicator values to calculate integrated scores.
}
\keyword{datasets}
